create interface to manage main logic
